Parts present, debugging begins

+++ database/databaseSupport.h
@@ -7,7 +7,7 @@ double GetDLOD (int, double,
 		double, double, double, double,
 		int);
 void SignOn (char *, int, char *, int, char *);
-int GetDWork (int, int, double *, char *, double *,
+int GetDWork (double, double, double *, char *, double *,
 	      double *, double *, double *, double *,
 	      double *, double *, double *, double *,
 	      double *, double *, double *, double *);
